Fix problem with error in native code
Previously there would be an error in native code after the application shutdown, causing a large and scary looking dump to be printed out. It is unclear what exactly causes this, but using the StartCapture() method instead of the Capture() method seems to fix it.

The use of StopCaptureTimeout is needed because without it the program hangs after terminating. Again, the cause is unclear, but even with a timeout of 1s, it shuts down immediately.